Move all pages to pages directory
