+++ client/src/pages/Home.js
@@ -8,8 +8,8 @@ import '../css_mobile/Home.css';
 
 import useWindowDimensions from '../utils/mobile';
 
-import { HomeBlock2Desktop, HomeBlock3Desktop } from './desktop/HomeDesktop';
-import { HomeBlock2Mobile, HomeBlock3Mobile } from './mobile/HomeMobile';
+import { HomeBlock2Desktop, HomeBlock3Desktop } from '../components/desktop/HomeDesktop';
+import { HomeBlock2Mobile, HomeBlock3Mobile } from '../components/mobile/HomeMobile';
 
 const Home = () => {
   const isMobile = useWindowDimensions()[1];
